A  *significant* refactor of the interfaces to reduce the amount of complexity involved in adding new tests.
As we get further down the sphinx4 road, the number of test setups is
begining to grow significantly.  We have  two sets of acoustic
models (tidigits, wsj), two linguists (Static and Tree), three sets of
test inputs, (TI46, TIDIGITS, AN4), with subsets for each (quick and
super_quick).

As I was experimenting with the AN4 test data and the new linguists
and new acoustic models, it was clear that, to keep the test
environment clear, easy to understand, and easy to add new tests
something would have to be done. To add a simple spelling test I had
to:    Add a new SpellingTreeLinguist class, to test with the
TreeLinguist, add a new SpellingStaticLinguist class to test with the
static linguist, add a new grammar text file, add two new sets of
property files, modfiy the tests.Decoder class,
add a new test input batch file and add two new
targets to the makefile. That is 8 separate files, clearly this was
crazy and something had to be done.

The first refactor (which I described in an earlier email) allows
grammars and linguists to be used in a mix-and-max fashion instead of
requiring subclasses.  This allows us to have a tidigits test using
the tree or static linguist without any code changes.

The second refactor was to restructure how linguist are constructed so
we can easily chose a linguist based upon a SphinxProperty. This means
that test drivers such as test.Decoder do not have to have specific
knowledge about the type of linguist.  Without this change, the test
driver would have to be modified/extended with every new linguist.

The third refactor was to create a "Recognizer' class that creates
and configures and connects up all of the parts of the recognizer (the
front end, acoustic model, linguist, grammar, search).  Previously
this work was done in test code.  Getting the recognizer to work at a
high level would require too much intimate knowledge of the structure
of sphinx 4.  The Recognizer is still not designed to be a full blown
sphinx api, it just cleans things up quite a bit, and moves the
central construction of a recognizer out of the test code and into the
main code body.

The final refactor was to reduce the number of files needed to support
our tests. There were a number of grammar and properties files in the
test/Decoder directory that were needed to run the various test
targets.  The need for many of these was eliminated. The 'quick' and
'super_quick' versions of the test input batch files were eliminated
by adding a 'skip' property to the BatchDecoder that allows us to use
a large batch file but skip over entries (i.e. use only every fifth
line out of the batch file).

Many of the properties files were identical except for one or two
parameters.  These were eliminated and make targets that set the
modified parameter from the javac command line are used instead.

There were a number of dusty and vestigial classes (old ways of
pruning, primitive ActiveLists and such). These have been cleaned out
as well.

Now to add a new test, a new grammar file and a new test input batch
file need to be added and a few new targets added to the makefile.
This is much easier.


git-svn-id: https://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@540 94700074-3cef-4d97-a70e-9c8c206c02f5