Updated regression tests to always check out (cvs update) the regression.log just before updating it, and checking it in (cvs commit) immediately afterward to minimize the chances that an update conflict occurs that will cause the commit of the regression.log.
This reduces the window of vulnerabity to a simultaneous update from 8
hours (the length of the regresion tests) to a couple of seconds.

The down side is that there will be commit of the regression.log for
every test that runs, which means that every morning we'll have
half-a-dozen regression commit emails to go with our breakfast spam.


git-svn-id: https://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@829 94700074-3cef-4d97-a70e-9c8c206c02f5