This is 'work-in-progress'. This is the reduced footprint version of the compressed lextree, The goal being to reduce the overall footprint of large vocabulary tasks (60K words).  This checkin is to protect the work so that is not lost. It is not complete yet. The unigram smear is not working properly yet. Also, it is not fully/properly documented.  Once the unigram smear is fixed this will replace the 'clextree' linguist.
Initial tests show that with this linguist we can run at the same speed
as the older linguist, but in a much smaller space. The heap reqs are
less than 500mb for this linguist while the onld clextree linguist
requires about 900mb. Also, the time to startup (not counting AM loading)
is reduced from 60 seconds to less than 10 seconds.


git-svn-id: https://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@2267 94700074-3cef-4d97-a70e-9c8c206c02f5