An upgrade to zipcity. ZipCity now shows a map of zip codes. When a zipcode is recognized, the zipcode is drawn on the map.
Note that this is not completely done. I wanted to check it in to make
sure that the changes were archived.

The zipcode database is derived from a different source and is a bit
smaller than the previous one, but is also more accurate for the lat
and long data (much of the lat/long data from the previous db was
wrong or missing).

I haven't completed the javadocs yet, nor have I exhaustively tested
this.  I will do this tomorrow.

Also, i need to add the endpointing version of zipcity.


git-svn-id: https://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@3488 94700074-3cef-4d97-a70e-9c8c206c02f5